requirements.txt
numpy
torch
dataclasses

# For reading ros bags
--extra-index-url https://rospypi.github.io/simple/
rospy
rosbag

# For reading pngs out of rosbags
pypng

# For loading and persisting data efficiently
pandas
pyarrow

# For running the same ONNX models as run on the bot locally, for loss function estimation
onnx>=1.10
onnxruntime-gpu

# For running non max supression and other utility ops
torchvision>=0.10.0


# For reporting and saving training logs
wandb

# console progress bar
tqdm

# For unit testing (QuickCheck style)
hypothesis

# For loading and saving audio files
soundfile
